move project health table files
