+++ src/component/project/Project/ProjectHealth/ReportTable/ReportExpiredCell/ReportExpiredCell.tsx
@@ -1,6 +1,6 @@
 import { VFC } from 'react';
 import { DateCell } from 'component/common/Table/cells/DateCell/DateCell';
-import { IReportTableRow } from 'component/Reporting/ReportTable/ReportTable';
+import { IReportTableRow } from 'component/project/Project/ProjectHealth/ReportTable/ReportTable';
 import { TextCell } from 'component/common/Table/cells/TextCell/TextCell';
 
 interface IReportExpiredCellProps {
